requirements.txt
# Usage: pip install -r requirements.txt

# Base ----------------------------------------

matlabengineforpython # Matlab engine for python